pyproject.toml
[project]
name = "zufang-cli"
version = "0.1.0"
description = "CLI for aggregating public rental listings from Anjuke and Beike"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Codex" }]
dependencies = [
    "beautifulsoup4>=4.12",
    "click>=8.1",
    "httpx>=0.27",
    "PyYAML>=6.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
]

[project.scripts]
zufang = "zufang_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["zufang_cli"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
